requirements.txt
phidata
openai
wikipedia-api
